+++ test/features/device_sync/sharing_settings_section_test.dart
@@ -31,7 +31,7 @@ class _FakeSharingService implements LocalSharingService {
     pendingItems: 0,
     isTrusted: false,
     isBlocked: false,
-    appVersion: '1.1.2',
+    appVersion: '1.1.3',
     protocolVersion: '1',
   );